+++ sdk/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [0.3.4](https://github.com/opentdf/platform/compare/sdk/v0.3.3...sdk/v0.3.4) (2024-07-23)
+
+
+### Bug Fixes
+
+* policy binding fix ([#1198](https://github.com/opentdf/platform/issues/1198)) ([6bf8e74](https://github.com/opentdf/platform/commit/6bf8e747885c05ea6a23db707e778b16239abe0a))
+
+## [0.3.3](https://github.com/opentdf/platform/compare/sdk/v0.3.2...sdk/v0.3.3) (2024-07-22)
+
+
+### Bug Fixes
+
+* fixed policy binding type ([#1184](https://github.com/opentdf/platform/issues/1184)) ([9800a32](https://github.com/opentdf/platform/commit/9800a32c8d9d83458403e2f87720f7882461fc32))
+* **sdk:** Allow empty kas info list ([#1161](https://github.com/opentdf/platform/issues/1161)) ([dd6db8e](https://github.com/opentdf/platform/commit/dd6db8e370142be647aba12f00f466ea6d680297))
+* **sdk:** Remove case sensitivity of attr values ([#1160](https://github.com/opentdf/platform/issues/1160)) ([21d73f6](https://github.com/opentdf/platform/commit/21d73f6b6af88ecdfeb17c1db3fbfbb88cde89b5))
+
 ## [0.3.2](https://github.com/opentdf/platform/compare/sdk/v0.3.1...sdk/v0.3.2) (2024-07-14)